+++ Ranger/Services/PokerTracker/Model/BasicPlayerStatistics/BasicPlayerStatisticsQuery.swift
@@ -1,31 +1,26 @@
 //
-//  BasicPlayerStatisticsCollection.swift
+//  LiveTourneyTableQuery.swift
 //  Ranger
 //
-//  Created by Geri Borbás on 2019. 12. 09..
+//  Created by Geri Borbás on 2019. 12. 14..
 //  Copyright © 2019. Geri Borbás. All rights reserved.
 //
 
 import Foundation
-import PostgresClientKit
 
 
-class BasicPlayerStatisticsCollection: Collection
+struct BasicPlayerStatisticsQuery: Query
 {
     
     
-    typealias RowType = BasicPlayerStatistics
-    var rows: [BasicPlayerStatistics]
+    typealias EntryType = BasicPlayerStatistics
     
-    var queryString: String
+    
+    var string: String
     {
         let queryFilePath = Bundle.main.path(forResource: "BasicPlayerStatisticsQuery", ofType: "sql")
         guard let queryString = try? String(contentsOfFile: queryFilePath!, encoding: String.Encoding.utf8)
         else { return "" }
         return queryString
     }
-    
-    
-    init()
-    { rows = [] }
 }