+++ Ranger/Services/PokerTracker/Model/Entry.swift
@@ -10,7 +10,7 @@ import Foundation
 import PostgresClientKit
 
 
-protocol RowInitable: CustomStringConvertible
+protocol Entry: CustomStringConvertible
 {